user and transaction classes moved to separate folder
