+++ lib/presentation/widgets/user_table.dart
@@ -8,7 +8,7 @@ import 'package:meta_app/presentation/widgets/colored_button.dart';
 import 'package:meta_app/presentation/widgets/editing_field.dart';
 import 'package:meta_app/presentation/widgets/responsive.dart';
 
-import '../pages/client_profile/sections/dashboard_tab.dart';
+import '../../data/models/user.dart';
 
 class UserTable extends StatelessWidget {
   final List<User> users;
@@ -395,24 +395,4 @@ class _ShowTextState extends State<_ShowText> {
       ),
     );
   }
-}
-
-class User {
-  final String name;
-  final String email;
-  final int referralLevel;
-  final double availableBalance;
-  final List<Transaction> pendingTransactions;
-  final List<Transaction> transactionsHistory;
-  final List<User> referrals;
-
-  const User({
-    required this.name,
-    required this.email,
-    required this.referralLevel,
-    required this.availableBalance,
-    required this.pendingTransactions,
-    required this.transactionsHistory,
-    required this.referrals,
-  });
 }
